backend/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
groq==0.11.0
pinecone-client==5.0.1
protoc-gen-openapiv2==0.0.1
python-dotenv==1.0.1